Added layer thickness prediction
